+++ src/main/java/gg/moonflower/etched/core/fabric/EtchedFabricClient.java
@@ -4,7 +4,7 @@ import gg.moonflower.etched.common.network.EtchedMessages;
 import gg.moonflower.etched.core.Etched;
 import net.fabricmc.api.ClientModInitializer;
 
-//import gg.moonflower.etched.common.entity.MinecartJukebox;
+import gg.moonflower.etched.common.entity.MinecartJukebox;
 import gg.moonflower.etched.core.EtchedClient;
 import net.fabricmc.fabric.api.event.client.player.ClientPickBlockGatherCallback;
 import net.minecraft.client.resources.model.ModelResourceLocation;
@@ -26,15 +26,15 @@ public class EtchedFabricClient implements ClientModInitializer {
         EtchedClient.registerRenderers();
 
         ClientPickBlockGatherCallback.EVENT.register((player, result) -> {
-            if (result.getType() == HitResult.Type.ENTITY && player.getAbilities().instabuild) {
+            if (result.getType() == HitResult.Type.ENTITY) {
                 Entity entity = ((EntityHitResult) result).getEntity();
-                //FIXME
-                //if (entity instanceof MinecartJukebox minecart) {
-                //    return new ItemStack(minecart.getDropItem());
-                //}
+                if (entity instanceof MinecartJukebox minecart) {
+                    return new ItemStack(minecart.getDropItem());
+                }
             }
             return ItemStack.EMPTY;
         });
+        EtchedClient.registerItemProperties();
         EtchedClient.registerItemColors();
     }
 }